requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
python-dotenv==1.2.2
qrcode==8.2
requests==2.33.1
urllib3==2.6.3
